LPD-76850 Moving Bulk actions to CMS and CMP module
