+++ modules/apps/object/object-api/src/main/java/com/liferay/object/entry/folder/util/ObjectEntryFolderUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.entry.folder.util;
+package com.liferay.object.entry.folder.util;
 
 import com.liferay.object.constants.ObjectEntryFolderConstants;
 import com.liferay.object.model.ObjectEntryFolder;
