+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/bulk/selection/DueDateObjectBulkSelectionAction.java
@@ -3,20 +3,19 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.bulk.selection;
+package com.liferay.site.cms.site.initializer.internal.bulk.selection;
 
 import com.liferay.bulk.selection.BulkSelectionAction;
-import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.service.ObjectEntryService;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.util.HashMapBuilder;
+import com.liferay.site.cms.site.initializer.bulk.selection.BaseObjectBulkSelectionAction;
 
 import java.io.Serializable;
 
-import java.util.HashMap;
+import java.util.Date;
 import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
@@ -26,10 +25,10 @@ import org.osgi.service.component.annotations.Reference;
  * @author Danny Situ
  */
 @Component(
-	property = "bulk.selection.action.key=status.object",
+	property = "bulk.selection.action.key=due.date.object",
 	service = BulkSelectionAction.class
 )
-public class StatusObjectBulkSelectionAction
+public class DueDateObjectBulkSelectionAction
 	extends BaseObjectBulkSelectionAction {
 
 	@Override
@@ -41,30 +40,20 @@ public class StatusObjectBulkSelectionAction
 			return;
 		}
 
-		String status = (String)inputMap.get("status");
+		Date dueDate = (Date)inputMap.get("dueDate");
 
-		if (Validator.isBlank(status)) {
+		if (dueDate == null) {
 			return;
 		}
 
 		ObjectEntry objectEntry = (ObjectEntry)object;
 
-		Map<String, Serializable> properties = new HashMap<>();
-
-		ObjectDefinition objectDefinition =
-			objectDefinitionLocalService.getObjectDefinition(
-				objectEntry.getObjectDefinitionId());
-
-		if (StringUtil.equals(objectDefinition.getName(), "CMPTask")) {
-			properties.put("state", status);
-		}
-		else {
-			properties.put("status", status);
-		}
-
 		_objectEntryService.partialUpdateObjectEntry(
 			objectEntry.getObjectEntryId(),
-			objectEntry.getObjectEntryFolderId(), properties,
+			objectEntry.getObjectEntryFolderId(),
+			HashMapBuilder.<String, Serializable>put(
+				"dueDate", dueDate
+			).build(),
 			new ServiceContext());
 	}
