+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/bulk/selection/StatusObjectBulkSelectionAction.java
@@ -3,18 +3,21 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.bulk.selection;
+package com.liferay.site.cms.site.initializer.internal.bulk.selection;
 
 import com.liferay.bulk.selection.BulkSelectionAction;
+import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.service.ObjectEntryService;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.util.HashMapBuilder;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.site.cms.site.initializer.bulk.selection.BaseObjectBulkSelectionAction;
 
 import java.io.Serializable;
 
-import java.util.Date;
+import java.util.HashMap;
 import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
@@ -24,10 +27,10 @@ import org.osgi.service.component.annotations.Reference;
  * @author Danny Situ
  */
 @Component(
-	property = "bulk.selection.action.key=due.date.object",
+	property = "bulk.selection.action.key=status.object",
 	service = BulkSelectionAction.class
 )
-public class DueDateObjectBulkSelectionAction
+public class StatusObjectBulkSelectionAction
 	extends BaseObjectBulkSelectionAction {
 
 	@Override
@@ -39,20 +42,30 @@ public class DueDateObjectBulkSelectionAction
 			return;
 		}
 
-		Date dueDate = (Date)inputMap.get("dueDate");
+		String status = (String)inputMap.get("status");
 
-		if (dueDate == null) {
+		if (Validator.isBlank(status)) {
 			return;
 		}
 
 		ObjectEntry objectEntry = (ObjectEntry)object;
 
+		Map<String, Serializable> properties = new HashMap<>();
+
+		ObjectDefinition objectDefinition =
+			objectDefinitionLocalService.getObjectDefinition(
+				objectEntry.getObjectDefinitionId());
+
+		if (StringUtil.equals(objectDefinition.getName(), "CMPTask")) {
+			properties.put("state", status);
+		}
+		else {
+			properties.put("status", status);
+		}
+
 		_objectEntryService.partialUpdateObjectEntry(
 			objectEntry.getObjectEntryId(),
-			objectEntry.getObjectEntryFolderId(),
-			HashMapBuilder.<String, Serializable>put(
-				"dueDate", dueDate
-			).build(),
+			objectEntry.getObjectEntryFolderId(), properties,
 			new ServiceContext());
 	}
